Added player health and cubicle boundaries
Changed the player rigidbody to dynamic with gravity set to 0 and rotation on the z axis off. Cubicle has a polygon collider with a kinematic rigidbody. Added a capsule collider to the light with istrigger on. Player health will go down every 0.1 seconds if colliding with the light. The capsule collider moves with the light.